+++ src/NadekoBot/Migrations/NadekoContextModelSnapshot.cs
@@ -245,6 +245,34 @@ namespace NadekoBot.Migrations
                     b.ToTable("AutoPublishChannel");
                 });
 
+            modelBuilder.Entity("NadekoBot.Db.Models.AutoThreadChannel", b =>
+                {
+                    b.Property<int>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("INTEGER");
+
+                    b.Property<int>("ArchiveDurationMinutes")
+                        .HasColumnType("INTEGER");
+
+                    b.Property<ulong>("ChannelId")
+                        .HasColumnType("INTEGER");
+
+                    b.Property<ulong>("GuildId")
+                        .HasColumnType("INTEGER");
+
+                    b.Property<int>("Mode")
+                        .HasColumnType("INTEGER");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("ChannelId")
+                        .IsUnique();
+
+                    b.HasIndex("GuildId");
+
+                    b.ToTable("AutoThreadChannel");
+                });
+
             modelBuilder.Entity("NadekoBot.Db.Models.AutoTranslateChannel", b =>
                 {
                     b.Property<int>("Id")